refacto(launchers): move auto doc at filesyntax level
we had the filesynatx documentation for launchers at their python API level which was not correct.

We simplify the launcher objects as not being aware they are used in a file format and move that documentation in a dedicated module.

we add quick assertion tests to ensure the developer did not added a new launcher without adding its doc