!!refacto: complex change of serialization system
create a proper separation in the code between the kloch core python API and it's serialization.

We also introduce a new "merge_system_environ" field for additional flexibility.

We move the automated launcher documentation again by integrating it in the serialized objects. "The documentation validation" step is however now performed in unittests instead of at runtime.

To retrieve launchers we need to manually define them into a list instead of using the less predictable `BaseLauncher.__subclasses__()`